requirements.txt
bcrypt==4.3.0
pillow==11.2.1
